moved category helpers into a sub directory
