+++ app/lib/handlers/information_intent_request_handler.rb
@@ -10,7 +10,7 @@ module Handlers
 
     def get_intent_specific_method(intent_name)
       {
-        'GetExpiration' => :get_expiration_details
+        'GetExpiration' => :process_get_expiration
       }[intent_name]
     end
 
@@ -18,13 +18,19 @@ module Handlers
       @item ||= intent['slots']['item']['value']
     end
 
-    def get_expiration_details
-      details = Helpers::CategoryHelper.categorize item
-      if details[:small_category] == :other
+    def item_details
+      @item_details ||= Helpers::CategoryHelper.categorize item
+    end
+
+    def message
+      if item_details[:small_category] == :other
         message = "Unable to find expiry information for #{item}"
       else
         message = "When appropriately stored, #{item} is good for #{details[:time_til_expiration]} days"
       end
+    end
+
+    def process_get_expiration
       Helpers::HandlerHelper.create_response(
         message: message,
         card: {
@@ -36,7 +42,7 @@ module Handlers
         end_session: false
       )
     end
-
+    
   end
 
 end
