+++ app/lib/helpers/category_helpers/other_category_helper.rb
@@ -1,0 +1,16 @@
+module Helpers
+  module CategoryHelpers
+
+    class OtherCategoryHelper < SubcategoryHelperBase
+
+      CATEGORIES = {
+        other: {
+          warning: 0,
+          expiring: 0
+        }
+      }
+
+    end
+
+  end
+end
