--- app/lib/helpers/other_category_helper.rb
@@ -1,14 +1,0 @@
-module Helpers
-
-  class OtherCategoryHelper < SubcategoryHelper
-
-    CATEGORIES = {
-      other: {
-        warning: 0,
-        expiring: 0
-      }
-    }
-
-  end
-
-end
