+++ spec/lib/pantry/helpers/dairy_category_helper_spec.rb
@@ -1,16 +1,16 @@
 require 'rails_helper'
 
-RSpec.describe Helpers::DairyCategoryHelper do
+RSpec.describe Helpers::CategoryHelpers::DairyCategoryHelper do
 
   describe ".categorize" do
     it "returns a category corresponding to an item name if one matches" do
-      category_name = Helpers::DairyCategoryHelper.categorize 'milk'
+      category_name = Helpers::CategoryHelpers::DairyCategoryHelper.categorize 'milk'
 
       expect(category_name).to eq :milk
     end
 
     it "returns the :other category if the item name matches with no other categories" do
-      category_name = Helpers::DairyCategoryHelper.categorize 'meat'
+      category_name = Helpers::CategoryHelpers::DairyCategoryHelper.categorize 'meat'
 
       expect(category_name).to eq :other
     end
@@ -18,7 +18,7 @@ RSpec.describe Helpers::DairyCategoryHelper do
 
   describe ".time_til_warning" do
     it "returns the amount of time in days until an item in a given category creates an expiration warning" do
-      time = Helpers::DairyCategoryHelper.time_til_warning(category: :milk)
+      time = Helpers::CategoryHelpers::DairyCategoryHelper.time_til_warning(category: :milk)
 
       expect(time).to eq 5
     end
@@ -26,7 +26,7 @@ RSpec.describe Helpers::DairyCategoryHelper do
 
   describe ".time_til_expiration" do
     it "returns the amount of time in days until an item in a given category expires" do
-      time = Helpers::DairyCategoryHelper.time_til_expiration(category: :milk)
+      time = Helpers::CategoryHelpers::DairyCategoryHelper.time_til_expiration(category: :milk)
 
       expect(time).to eq 7
     end
@@ -34,7 +34,7 @@ RSpec.describe Helpers::DairyCategoryHelper do
 
   describe ".allWords" do
     it "returns all words associated with dairy in a set" do
-      allWords = Helpers::DairyCategoryHelper.all_words
+      allWords = Helpers::CategoryHelpers::DairyCategoryHelper.all_words
 
       expect(allWords).to include('1-percent')
       expect(allWords).to include('whipping')
